Added Daily Report class, made a print method for it

+++ src/OrderID.java
@@ -1,16 +1,16 @@
 import java.util.Random;
 
 public class OrderID {
-    private final Long randomID;
+    private final Integer randomID;
     private final String hexStringID;
 
     OrderID() {
         this.randomID = generateRandomID();
-        this.hexStringID = Long.toHexString(randomID);
+        this.hexStringID = Integer.toHexString(randomID);
     }
 
-    private Long generateRandomID() {
-        return Math.abs(new Random().nextLong());
+    private Integer generateRandomID() {
+        return Math.abs(new Random().nextInt());
     }
 
     public String getHexStringID() {
